Create Service Classes to make mapping cleaner
